+++ ogma-core/templates/copilot-cfs/fsw/src/copilot_cfs.c
@@ -173,7 +173,7 @@ void COPILOT_Process{{msgDataDesc}}(void)
     {{/msgDataFromField}}
 
     // Run all copilot monitors.
-    step();
+    copilot_step();
 }
 
 {{/msgHandlers}}